// src/api/generated/models/media-asset-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ankieter API
 * Ankieter 6.0 - the API contract, source of truth for the backend and the frontend
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { MediaKindDto } from './media-kind-dto';

/**
 * 
 * @export
 * @interface MediaAssetDto
 */
export interface MediaAssetDto {
    /**
     * Read the file at /api/v1/media/{id}.
     * @type {string}
     * @memberof MediaAssetDto
     */
    'id': string;
    /**
     * 
     * @type {MediaKindDto}
     * @memberof MediaAssetDto
     */
    'kind': MediaKindDto;
    /**
     * The language the material is for. Absent when it serves every language: a photo of a shelf does, a film with a Polish voice-over does not.
     * @type {string}
     * @memberof MediaAssetDto
     */
    'locale'?: string;
    /**
     * What the author\'s computer called the file. For people only.
     * @type {string}
     * @memberof MediaAssetDto
     */
    'filename': string;
    /**
     * Read from the content, not from the name.
     * @type {string}
     * @memberof MediaAssetDto
     */
    'mimeType': string;
    /**
     * 
     * @type {number}
     * @memberof MediaAssetDto
     */
    'sizeBytes': number;
    /**
     * As shown, a photo or film taken upright on a phone included.
     * @type {number}
     * @memberof MediaAssetDto
     */
    'width': number;
    /**
     * 
     * @type {number}
     * @memberof MediaAssetDto
     */
    'height': number;
    /**
     * VIDEO only. Counts towards the time a test allows.
     * @type {number}
     * @memberof MediaAssetDto
     */
    'durationMs'?: number;
    /**
     * 
     * @type {string}
     * @memberof MediaAssetDto
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof MediaAssetDto
     */
    'createdBy': string;
}